requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
djangorestframework==3.11.0
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
Keras==2.4.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
